Integer --> BigInteger로 변경

+++ src/main/java/com/basquiat/address/node/BlockChainNodeInterface.java
@@ -1,5 +1,6 @@
 package com.basquiat.address.node;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -20,14 +21,14 @@ public interface BlockChainNodeInterface {
 	
 	public String getNewAddress() throws Exception;
 	
-	public Integer getBlockCount() throws Exception;
+	public BigInteger getBlockCount() throws Exception;
 	
-	public String getBlockHash(Integer blockNumber) throws Exception;
+	public String getBlockHash(BigInteger blockNumber) throws Exception;
 	
 	public JsonObject getBlock(String blockHash) throws Exception;
 	
 	public JsonObject getRawTransaction(String txId) throws Exception;
 
-	public Map<String, Object> schedulingTransactionCheck(Integer lastBlock) throws Exception;
+	public Map<String, Object> schedulingTransactionCheck(BigInteger lastBlock) throws Exception;
 	
 }
